Split task tests into driftbandit and twostep.
